Add files via upload
Implement JWT authentication

- Add Spring Security with JWT-based authentication and authorization
- Create JWTService for token generation/validation and JWTFilter for request interception
- Implement UserDetailsService (MyUserDetailService) and UserPrincipal for custom user details
-Configure security endpoints to permit login/register and Swagger docs, requiring authentication for others

+++ FinanceTracker/src/main/java/com/example/FinanceTracker/Service/UserService.java
@@ -1,8 +1,12 @@
 package com.example.FinanceTracker.Service;
 
-import com.example.FinanceTracker.Model.User;
+import com.example.FinanceTracker.Model.Users;
 import com.example.FinanceTracker.Repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.authentication.AuthenticationManager;
+import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
 import java.util.List;
 
@@ -10,29 +14,58 @@ import java.util.List;
 public class UserService {
 
     @Autowired
+    private JWTService jwtService;
+
+    @Autowired
+    private AuthenticationManager authmanager;
+
+    @Autowired
     private UserRepository Repo;
 
-    public List<User>
+    private BCryptPasswordEncoder encoder = new BCryptPasswordEncoder(12);
+
+    public List<Users>
     getAllUsers() {
         return Repo.findAll();
     }
 
-    public User
+    public Users
     getUserById(Long id) {
         return Repo.findById(id).orElse(null);
     }
 
-    public User
-    createUser(User user) {
-        return Repo.save(user);
+    public Users
+    createUser(Users users) {
+        users.setPassword(encoder.encode(users.getPassword()));
+        users.setRole("ROLE_USER"); // Set a default role
+        return Repo.save(users);
     }
 
     public boolean
-    deleteUser(Long id) {
-        if (Repo.existsById(id)) {
+    deleteUser(Long id, String email,String password) {
+        if (Repo.existsById(id) && encoder.matches(password,getUserById(id).getPassword())) {
             Repo.deleteById(id);
             return true;
         }
         return false;
     }
+
+
+    public String verify(Users user) {
+        try {
+            Authentication authentication = authmanager.authenticate(
+                    new UsernamePasswordAuthenticationToken(user.getUsername(), user.getPassword())
+            );
+
+            if (authentication.isAuthenticated()) {
+                return jwtService.generateToken(user.getUsername());
+            }
+        } catch (Exception e) {
+            return "Fail: " + e.getMessage();
+        }
+        return "Fail: Not Authenticated";
+    }
+
+
+
 }
